+++ R/pdb.R
@@ -1,0 +1,8 @@
+#' Planning Database (Published in year 2015)
+#' @description
+#' The Planning Database (pdb) contains selected data from the 2010 Census and the 2009-2013, 5-years American Community Survey (ACS) estimates.
+#' @format A dataset with 815 observations and 344 attributes.
+#' @usage data(pdb)
+#' @source \url{https://www.census.gov/data/datasets/2015/adrm/research/2015-planning-database.html}
+"pdb"
+#> [1] "pdb"