+++ R/prostate.R
@@ -1,17 +1,17 @@
 #' Prostate Levels
-#' @description 
-#' The data describe the level of a prostate-specific antigen 
-#' associated with eight clinical measures in 97 male patients 
-#' taking a radical prostatectomy.
-#' 
+#' @description
+#' The dataset comprises the level of prostate-specific antigen
+#' associated with eight clinical measures in 97 male patients
+#' undergoing a radical prostatectomy.
+#'
 #' @usage data(prostate)
 #'
 #'
-#' @references 
+#' @references
 #' Stamey, T. A., Kabalin, J. N., McNeal, J. E., Johnstone, I. M., Freiha, F., Redwine, E. A. et al.
-#'(1989). Prostate specific antigen in the diagnosis and treatment of adenocarcinoma of the
-#'prostate. II. Radical prostatectomy treated patients. _The Journal of Urology_
-#'141, 1076, 1083.
+#' (1989). Prostate specific Antigen in the Diagnosis and Treatment of Adenocarcinoma of the
+#' Prostate. II. \emph{Radical prostatectomy treated patients. The Journal of Urology}.
+#' 141, 1076, 1083.
 #'
 "prostate"
 #> [1] "prostate"