+++ R/raman.R
@@ -1,0 +1,15 @@
+#' Raman Spectroscopy
+#'
+#' @description
+#' The Raman dataset contains 69 samples of providing fatty acid information in terms of
+#' the percentage of total sample weight and the percentage of total fat content
+#'
+#' @usage data(raman)
+#'
+#' @references
+#' Naes T., Tomic O., Afseth N.K., Segtnan V., Mage I. (2013)
+#' Multi-Block Regression Based on Combinations of Orthogonalisation, Pls-Regression and Canonical
+#' Correlation Analysis. \emph{Chemometrics and Intelligent Laboratory Systems}. 124, 32-42.
+#'
+"raman"
+#> [1] "raman"