+++ R/recumbent.R
@@ -1,0 +1,15 @@
+#' Recumbent Cows
+#' @description
+#' The recumbent dataset contains information on pregnant dairy cows that become
+#' recumbent, i.e., they lie down shortly before or after calving for unknown reasons. This condition
+#' can be severe and frequently leads to death of the cow. Clark, Henderson,
+#' Hoggard, Ellison and Young (1987) analyzed data collected at the Ruakura (N.Z.)
+#' Animal Health Laboratory on a sample of recumbent cows.
+#' @format A dataset with 9 columns and 435 rows.
+#' @usage data(recumbent)
+#' @source
+#' Clark, R. G., Henderson, H. V., Hoggard, G. K. Ellison, R. S. and Young, B. J. (1987).
+#' The Abiltiy of Biochemical and Haematolgical Tests to Predict Recovery in Periparturient
+#' Recumbent Cows. \emph{ NZ Veterinary Journal}. 35, 126-133.
+"recumbent"
+#> [1] "recumbent"